Added LangSelector component
